workflow editor changes on ci pipeline and name suggestion api
