+++ util/helper.go
@@ -20,8 +20,10 @@ package util
 import (
 	"fmt"
 	"go.uber.org/zap"
+	"math/rand"
 	"strconv"
 	"strings"
+	"unsafe"
 )
 
 func ContainsString(list []string, element string) bool {
@@ -70,4 +72,15 @@ func Close(c Closer, logger *zap.SugaredLogger) {
 	if err := c.Close(); err != nil {
 		logger.Warnf("failed to close %v: %v", c, err)
 	}
+}
+
+var alphabet = []byte("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ")
+
+func Generate(size int) string {
+	b := make([]byte, size)
+	rand.Read(b)
+	for i := 0; i < size; i++ {
+		b[i] = alphabet[b[i]/5]
+	}
+	return *(*string)(unsafe.Pointer(&b))
 }
